content/html/content/test/test_bug500885.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=500885
-->
<head>
  <title>Test for Bug 500885</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/mockObjects.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=500885">Mozilla Bug 500885</a>
<div>
  <input id="file" type="file" />
</div>
<script type="text/javascript">

netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
Cu.import("resource://gre/modules/XPCOMUtils.jsm");

function MockFilePicker() { };
MockFilePicker.prototype = {
  QueryInterface: XPCOMUtils.generateQI([Ci.nsIFilePicker]),
  init: function(aParent, aTitle, aMode) { },
  appendFilters: function(aFilterMask) { },
  appendFilter: function(aTitle, aFilter) { },
  defaultString: "",
  defaultExtension: "",
  filterIndex: 0,
  displayDirectory: null,
  file: null,
  get fileURL() {
    throw Cr.NS_ERROR_NOT_IMPLEMENTED;
  },
  get files() {
    throw Cr.NS_ERROR_NOT_IMPLEMENTED;
  },
  show: function MFP_show() {
    return Ci.nsIFilePicker.returnOK;
  }
};

var mockFilePickerRegisterer = 
  new MockObjectRegisterer("@mozilla.org/filepicker;1",MockFilePicker);

function test() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var wu = window.QueryInterface(Ci.nsIInterfaceRequestor)
                  .getInterface(Ci.nsIDOMWindowUtils);
                  

  mockFilePickerRegisterer.register();
  try {
    var domActivateEvents;
    var fileInput = document.getElementById("file");
    var rect = fileInput.getBoundingClientRect();

    fileInput.addEventListener ("DOMActivate", function () {
      domActivateEvents++;
    }, false);

    domActivateEvents = 0;
    wu.sendMouseEvent("mousedown", rect.left + 5, rect.top + 5, 0, 1, 0);
    wu.sendMouseEvent("mouseup", rect.left + 5, rect.top + 5, 0, 1, 0);
    is(domActivateEvents, 1, "click on text field should only fire 1 DOMActivate event");

    domActivateEvents = 0;
    wu.sendMouseEvent("mousedown", rect.right - 5, rect.top + 5, 0, 1, 0);
    wu.sendMouseEvent("mouseup", rect.right - 5, rect.top + 5, 0, 1, 0);
    is(domActivateEvents, 1, "click on button should only fire 1 DOMActivate event");

  } finally {
    SimpleTest.executeSoon(unregister);
    
  }
}

function unregister()
{
  mockFilePickerRegisterer.unregister();
  SimpleTest.finish();
}

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  setTimeout(test, 0);
};

</script>
</body>

</html>
